Move AuthenticationHelper to helper namespace
