+++ libraries/src/Joomla/Cms/Helper/AuthenticationHelper.php
@@ -6,7 +6,7 @@
  * @license    GNU General Public License version 2 or later; see LICENSE.txt
  */
 
-namespace Joomla\Cms\Authentication;
+namespace Joomla\Cms\Helper;
 
 defined('JPATH_PLATFORM') or die;
